Reworked how hold notes work.
a) The button must be held at the moment when the tick is passed. (Fixes #355)
b) The hold object must be pressed. (Fixes 'shield note' issues)
c) The last tick will be gracefully counted as long as the last button release is in the 'near window' of the end of the hold note (not 100% sure about this)